updateSystem only adm run

+++ public/view/updateSystem.php
@@ -1,9 +1,13 @@
 <?php
-if (file_exists(PATH_HOME . "_config/updates/version.txt"))
-    unlink(PATH_HOME . "_config/updates/version.txt");
 
-new \Config\UpdateSystem();
-?>
-<script>
-    location.href = "<?=HOME?>";
-</script>
+if ($_SESSION['userlogin']['id'] === "1") {
+    if (file_exists(PATH_HOME . "_config/updates/version.txt"))
+        unlink(PATH_HOME . "_config/updates/version.txt");
+
+    new \Config\UpdateSystem();
+    ?>
+    <script>
+        location.href = "<?=HOME?>";
+    </script>
+    <?php
+}
